+++ modules/kubernetes/datadog/variables.tf
@@ -34,17 +34,15 @@ variable "azure_config" {
   type = object({
     azure_key_vault_name = string
     identity = object({
-      client_id   = string
-      resource_id = string
-      tenant_id   = string
+      client_id = string
+      tenant_id = string
     })
   })
   default = {
     azure_key_vault_name = ""
     identity = {
-      client_id   = ""
-      resource_id = ""
-      tenant_id   = ""
+      client_id = ""
+      tenant_id = ""
     }
   }
 }